backend/hsk_asset_service/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
pillow
requests
supabase
python-dotenv
